Move keys and their version definition to the respective version modules
